requirements.txt
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-groq==0.0.1
faiss-cpu==1.7.4
groq-sdk==0.4.2
pymupdf==1.23.8
python-dotenv==1.0.0
whisper==1.1.10
sounddevice==0.4.6
numpy==1.24.3
pandas==2.0.3
tiktoken==0.5.1
sentence-transformers==2.2.2
streamlit-option-menu==0.3.6
streamlit-chat==0.1.1
protobuf==3.20.3
google-protobuf==3.20.3
